refactor: remove sveltekit; add layer.html entry
