+++ package.json
@@ -7,8 +7,6 @@
     "dev": "vite dev",
     "build": "vite build",
     "preview": "vite preview",
-    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
     "tauri": "tauri"
   },
   "license": "MIT",
@@ -22,8 +20,6 @@
     "@tauri-apps/plugin-updater": "~2"
   },
   "devDependencies": {
-    "@sveltejs/adapter-static": "^3.0.5",
-    "@sveltejs/kit": "^2.7.0",
     "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tailwindcss/typography": "^0.5.15",
     "@tauri-apps/cli": "^2.1.0",
@@ -41,11 +37,9 @@
   },
   "pnpm": {
     "ignoredBuiltDependencies": [
-      "@sveltejs/kit",
       "esbuild"
     ],
     "onlyBuiltDependencies": [
-      "@sveltejs/kit",
       "esbuild"
     ]
   }
