+++ src/pages/layer/index.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+  import { invoke } from "@tauri-apps/api/core";
+
+  $effect(() => {
+    console.log("🚀 ~ file: about.svelte:10 ~ getVersion ~ res:");
+  });
+</script>
+
+<div class="flex gap-2 flex-col">
+  hahahahahh
+</div>
