+++ src/pages/layer/index.ts
@@ -1,0 +1,6 @@
+import { mount } from 'svelte';
+import App from './index.svelte'
+
+const app = mount(App, { target: document.getElementById("app") as HTMLElement });
+
+export default app;
