--- src/routes/+layout.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-  import '../../node_modules/flowbite/dist/flowbite.min.css';
-	import '../app.css';
-
-	let { children } = $props();
-</script>
-
-{@render children()}
