+++ src/typing.d.ts
@@ -1,16 +1,14 @@
 // See https://svelte.dev/docs/kit/types#app.d.ts
 // for information about these interfaces
 declare global {
-  namespace App {
-    // interface Error {}
-    // interface Locals {}
-    // interface PageData {}
-    // interface PageState {}
-    // interface Platform {}
-    interface UserConfig {
-      auto_shuffle: boolean;
-      interval: number;
-    }
+  // interface Error {}
+  // interface Locals {}
+  // interface PageData {}
+  // interface PageState {}
+  // interface Platform {}
+  interface UserConfig {
+    auto_shuffle: boolean;
+    interval: number;
   }
   type BingImage = {
     copyrights: string[];
