+++ underlayer.html
@@ -2,11 +2,13 @@
 <html lang="en">
 	<head>
 		<meta charset="utf-8" />
-		<link rel="icon" href="%sveltekit.assets%/icon.png" />
+		<link rel="icon" href="public/icon.png" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		%sveltekit.head%
 	</head>
-	<body data-sveltekit-preload-data="hover">
-		<div id="app">%sveltekit.body%</div>
+	<body>
+		<div id="app">
+
+    </div>
+    <script type="module" src="./src/pages/layer/index.ts"></script>
 	</body>
 </html>
